Add Template Function for Game Score/Logic
Made a template function that should dictate the game scoring & logic.
(but currently the game logic is still dictated by numbers, when it should be dictated by the logic that I wrote).
I'm thinking of changing the values with a, b, c instead of 1, 2, 3 (Maybe that's the answer?).
So this commit is a great backup if things do go ape shit after I change the values!

I also still need to find answers how to make the game loop from round to round & then stop/restart after someone gets
3 points. Also how to keep up the score from round to round without the score deleting after a new round starts
or how to delete scores completely after the game is done / new game begins.

Lot of quetions, not so many answers!