Modify Number Choices with Words Instead
Tryed to change the game logic that I made with numbers to words instead (with no success).
I don't think it works, because the computer has to randomize a number instead of a string & when I tryed to
covert those numbers into strings the result is worse than it was before so...
(& I don't know how I could make it randomize between strings instead of numbers ether).

...I'm thinking going back to my last commit & keep this one on a seperate branch if I ever feel like giving this
path another go in the future!